refactor: deleted duplicate/ legacy tests and moved all files under backend/src/tests
