+++ backend/package.json
@@ -14,13 +14,13 @@
     "start:backtest": "node dist/runtime/backtest.js",
     "start:replay": "node dist/runtime/replay.js",
     "lint": "eslint src --ext .ts",
-    "test": "jest test/unit test/int test/http",
+    "test": "jest",
     "test:watch": "jest --watch",
-    "test:unit": "jest test/unit",
-    "test:int": "jest test/int",
-    "test:http": "jest test/http",
+    "test:unit": "jest src/tests/core src/tests/repositories src/tests/utils src/tests/adapters src/tests/strategy src/tests/services src/tests/controllers",
+    "test:int": "jest src/tests/int",
+    "test:http": "jest src/tests/http",
     "test:ci": "jest --coverage --ci",
-    "test:live": "RUN_INTEGRATION=1 jest test/live",
+    "test:live": "RUN_INTEGRATION=1 jest src/tests/live",
     "verify:http": "ts-node src/runtime/verify/httpSmoke.ts",
     "verify:backtest": "ts-node src/runtime/verify/backtestVerify.ts"
   },
