+++ backend/src/tests/core/computeMetrics.test.ts
@@ -1,8 +1,8 @@
 
-import { BacktestEngine } from "../../src/core/backtest/backtestEngine";
-import { Fill } from "../../src/types/orders";
-import { PortfolioSnapshot } from "../../src/types/portfolio";
-import { UUID } from "../../src/types/common";
+import { BacktestEngine } from "../../core/backtest/backtestEngine";
+import { Fill } from "../../types/orders";
+import { PortfolioSnapshot } from "../../types/portfolio";
+import { UUID } from "../../types/common";
 
 describe("BacktestEngine._computeMetrics", () => {
   let engine: any;
@@ -69,22 +69,9 @@ describe("BacktestEngine._computeMetrics", () => {
     const ec = [createSnapshot(100000), createSnapshot(100100)];
     const metrics = engine._computeMetrics(ec, fills, 100000, 0, 1000);
 
-    // Current implementation uses simple FIFO shift/push. 
-    // Two sells against one buy in FIFO map might count as 1.5 or fail if qty doesn't match.
-    // In our implementation: 
-    // buyMap.get(SPY) has [buy10]. 
-    // Sell 5 comes -> buys.shift() returns buy10. pnl calculated for 5?
-    // Wait, let's look at the implementation again.
-    /*
-          const matchedBuy = buys.shift()!;
-          const pnl = (fill.price - matchedBuy.price) * fill.qty ...
-    */
-    // If fill.qty (5) is less than matchedBuy.qty (10), it correctly calculates pnl for 5.
-    // BUT it shifted the whole buy10 out! So the second sell will find NO buy in the map.
-    // This is a known limitation of the simple FIFO matcher, but the goal is to test it.
-    
-    // In the current implementation, it would count 1 trade and leave the second sell unmatched.
-    expect(metrics.totalTrades).toBe(1); 
+    // Simple FIFO matcher shifts the entire buy on the first sell, leaving the second sell
+    // unmatched. Known limitation: partial closes count as 1 trade, not 2.
+    expect(metrics.totalTrades).toBe(1);
   });
 
   test("totalReturn is sourced from ledger (equity curve), not fills sum", () => {
