+++ backend/src/tests/int/engine.test.ts
@@ -1,12 +1,12 @@
 
-import { BacktestEngine } from "../../src/core/backtest/backtestEngine";
-import { BacktestLoader } from "../../src/core/backtest/backtestLoader";
-import { PairsStrategy } from "../../src/strategies/pairs/pairsStrategy";
-import { Bar } from "../../src/types/market";
-import { BacktestConfig } from "../../src/types/backtest";
-import { UUID } from "../../src/types/common";
+import { BacktestEngine } from "../../core/backtest/backtestEngine";
+import { BacktestLoader } from "../../core/backtest/backtestLoader";
+import { PairsStrategy } from "../../strategies/pairs/pairsStrategy";
+import { Bar } from "../../types/market";
+import { BacktestConfig } from "../../types/backtest";
+import { UUID } from "../../types/common";
 
-jest.mock("../../src/core/backtest/backtestLoader");
+jest.mock("../../core/backtest/backtestLoader");
 
 describe("BacktestEngine Integration", () => {
   let engine: BacktestEngine;
@@ -36,7 +36,7 @@ describe("BacktestEngine Integration", () => {
     const pattern = [0, 1, 2, 3, 2, 1, 0, -1, -2, -3, -2, -1, 0];
     const barsSPY = createSyntheticBars("SPY", 400, pattern.length, pattern);
     const barsQQQ = createSyntheticBars("QQQ", 300, pattern.length, pattern.map(v => -v)); // Opposite pattern for spread widening
-    
+
     const allBars = [...barsSPY, ...barsQQQ].sort((a, b) => a.ts - b.ts);
     (BacktestLoader.prototype.loadBars as jest.Mock).mockResolvedValue(allBars);
 
@@ -76,13 +76,13 @@ describe("BacktestEngine Integration", () => {
     process.env.BACKTEST_DEBUG = "1";
 
     const strategyFactory = () => [new PairsStrategy(config.strategyConfig as any)];
-    
+
     const result = await engine.run(config, strategyFactory);
 
     // Assert CHECK 7 invariants
     const fp = result.final_portfolio;
     const m = result.metrics;
-    
+
     const isClose = (a: number, b: number) => Math.abs(a - b) <= Math.max(1e-6, 1e-9 * Math.max(Math.abs(a), Math.abs(b)));
 
     expect(isClose(fp.equity, fp.cash + fp.positionsValue)).toBe(true);
@@ -93,7 +93,7 @@ describe("BacktestEngine Integration", () => {
     expect(m.maxDrawdown).toBeLessThanOrEqual(1);
     expect(result.equity_curve.length).toBeLessThanOrEqual(5000);
     expect(result.equity_curve[0].equity).toBe(INITIAL_CAPITAL);
-    
+
     expect(result.orders.length).toBeGreaterThan(0);
     expect(result.fills.length).toBeGreaterThan(0);
   });
@@ -145,8 +145,8 @@ describe("BacktestEngine Integration", () => {
 
     const config: any = {
         id: "run-3",
-        strategyConfig: { 
-            id: "strat-1", enabled: true, symbols: ["SPY", "QQQ"], leg1Symbol: "SPY", leg2Symbol: "QQQ", 
+        strategyConfig: {
+            id: "strat-1", enabled: true, symbols: ["SPY", "QQQ"], leg1Symbol: "SPY", leg2Symbol: "QQQ",
             rollingWindowMs: 3600000, olsWindowMs: 7200000, minObservations: 1,
             entryZScore: 1.0, exitZScore: 0.5, stopLossZScore: 10, maxHoldingTimeMs: 86400000,
             tradeNotionalUsd: 10000, priceSource: "mid", orderCooldownMs: 0,
@@ -157,7 +157,7 @@ describe("BacktestEngine Integration", () => {
         startDate: "2023-01-01T00:00:00Z", endDate: "2023-01-01T01:00:00Z"
     };
     const result = await engine.run(config as any, () => [new PairsStrategy(config.strategyConfig)]);
-    
+
     expect(result.final_portfolio.equity).toBeGreaterThan(0);
     // CHECK 7 holds
     const fp = result.final_portfolio;
@@ -174,8 +174,8 @@ describe("BacktestEngine Integration", () => {
 
     const config: any = {
         id: "run-4",
-        strategyConfig: { 
-            id: "strat-1", enabled: true, symbols: ["SPY", "QQQ"], leg1Symbol: "SPY", leg2Symbol: "QQQ", 
+        strategyConfig: {
+            id: "strat-1", enabled: true, symbols: ["SPY", "QQQ"], leg1Symbol: "SPY", leg2Symbol: "QQQ",
             rollingWindowMs: 3600000, olsWindowMs: 7200000, minObservations: 1,
             entryZScore: 1.0, exitZScore: 0.5, stopLossZScore: 10, maxHoldingTimeMs: 86400000,
             tradeNotionalUsd: 10000, priceSource: "mid", orderCooldownMs: 0,
@@ -185,39 +185,37 @@ describe("BacktestEngine Integration", () => {
         dataGranularity: "bar", slippageBps: 0, commissionPerShare: 0,
         startDate: "2023-01-01T00:00:00Z", endDate: "2023-01-01T01:00:00Z"
     };
-    
+
     const result = await engine.run(config as any, () => [new PairsStrategy(config.strategyConfig)]);
-    
+
     expect(result.orders.length).toBe(2);
-
-
   });
 
   test("INT 2.4 — Simulated clock isolation", async () => {
     const bars = createSyntheticBars("SPY", 100, 10, [0]);
     (BacktestLoader.prototype.loadBars as jest.Mock).mockResolvedValue(bars);
-    
+
     const config: any = { strategyConfig: { symbols: ["SPY"] }, initialCapital: 100000 };
     const strategyFactory = () => [];
-    
+
     const wallClockStart = Date.now();
     await engine.run(config, strategyFactory);
     const wallClockEnd = Date.now();
-    
+
     // Date.now() should be back to real time
     expect(Date.now()).toBeGreaterThanOrEqual(wallClockStart);
     expect(Date.now()).toBeLessThanOrEqual(wallClockEnd + 100); // 100ms buffer
-    
+
     // Test exception case
     const errorEngine = new BacktestEngine();
     (BacktestLoader.prototype.loadBars as jest.Mock).mockRejectedValue(new Error("Loader failed"));
-    
+
     try {
         await errorEngine.run(config, strategyFactory);
     } catch (e) {
         // Expected
     }
-    
+
     expect(Date.now()).toBeGreaterThanOrEqual(wallClockStart);
   });
 });
